Created data contracts ddb table in DL stack. Check if dc exists before moving file from raw to staged. Created data_contracts directory with json dcs. Python script runs after cdk deploy, to add json dcs to ddb table.

+++ data_lake/raw_to_staging.py
@@ -11,7 +11,22 @@ class RawToStaging:
         self.logger = logging.getLogger(self.__class__.__name__)
         self.logger.setLevel(logging.DEBUG)
         self.staging_bucket = os.getenv("STAGING_BUCKET_NAME")
+        self.data_contracts_table_name = os.getenv("DATA_CONTRACTS_TABLE_NAME")
         self.s3 = boto3.client('s3')
+        self.dynamodb = boto3.resource('dynamodb')
+
+    def data_contract_exists(self, key):
+        key_name = key.split('/')[0]
+        data_contracts_table = self.dynamodb.Table(self.data_contracts_table_name)
+        try:
+            self.logger.debug(f"Verifying {key_name=} in data contract table...")
+            response = data_contracts_table.get_item(
+                Key={"key_name": key_name}
+            )
+            return True
+        except Exception as e:
+            self.logger.warning(f"No data contract found for {key_name=}, with error {e}")
+            return False
 
     def copy_file_to_staged(self, key, bucket):
         self.logger.debug(f"Copying file with {key=} from {bucket} to {self.staging_bucket}...")
@@ -30,8 +45,9 @@ class RawToStaging:
             for record in json.loads(sqs_record["body"])["Records"]:
                 bucket = record["s3"]["bucket"]["name"]
                 key = urllib.parse.unquote_plus(record["s3"]["object"]["key"], encoding="UTF-8")
-                self.copy_file_to_staged(key, bucket)
-                self.delete_file_from_raw(key, bucket)
+                if self.data_contract_exists(key):
+                    self.copy_file_to_staged(key, bucket)
+                    self.delete_file_from_raw(key, bucket)
         self.logger.info("Finished lambda execution.")
         return "SUCCESS"
